switches to ESLint and wordpress validation (mostly)
