+++ .eslintrc.json
@@ -1,0 +1,12 @@
+{
+  "env": {
+    "es6": true,
+    "browser": true
+  },
+  "extends": [ "plugin:@wordpress/eslint-plugin/recommended" ],
+  "ignorePatterns": [ "gulpfile.js" ],
+  "rules": {
+    "quotes": ["error", "double"],
+    "no-console": [ "error", { "allow": ["warn", "error"] } ]
+  }
+}
