+++ package.json
@@ -20,22 +20,21 @@
   "license": "GPL-3.0",
   "licenseURI": "http://www.gnu.org/licenses/gpl.html",
   "devDependencies": {
+    "@wordpress/eslint-plugin": "^3.3.0",
     "autoprefixer": "^9.7.1",
     "gulp": "^4.0.2",
     "gulp-changed": "^4.0.2",
     "gulp-concat": ">=2.6.1",
+    "gulp-eslint": "^6.0.0",
     "gulp-header": "^2.0.9",
     "gulp-imagemin": "^6.1.1",
-    "gulp-jscs": ">=4.1.0",
-    "gulp-jshint": ">=2.0.4",
     "gulp-postcss": ">=8.0.0",
-    "gulp-rename": ">=1.4.0",
+    "gulp-rename": "^2.0.0",
     "gulp-sass": "^4.0.2",
     "gulp-shell": "^0.7.1",
     "gulp-sourcemaps": "^2.6.5",
     "gulp-strip-debug": ">=3.0.0",
-    "gulp-terser": "^1.2.0",
-    "jshint": "^2.10.3"
+    "gulp-terser": "^1.2.0"
   },
   "engines": {
     "node": ">=8.9.4",
