+++ src/sass/partials/_news.scss
@@ -1,8 +1,6 @@
 .news-post-detail {
 	margin: 2rem 0;
 	font-size: 1rem;
-	.contact {
-	}
 	.media-name, .media-phone {
 		margin-right: .5rem;
 	}
